[client]: Fix build problem caused by dependency incompatibility
Commit 4978f99a525e5ccb38f0ef7e6ec72d902d0168b0 added a new frontend
implemnetation which caused differen versions of react, react-dom and co. to be
added to the project's yarn.lock file. For some reason this broke the build of
the client project.

Upgrading the react and react-dom dependencies of the client package to the same
versions of those used by client-new seems to fix this issue. The rest of the
dependencies were also updated for good measure.