[client-new/tsconfig.json]: Extend the root tsconfig
For better consistency across the project. Note that this also changes the
`module` setting from `esnext` (the default for Next.js) to `commonjs`.

In the future, we would like to migrate the whole project to the es2020 module
system, but as it will require more work, we will postpone this for later.