[contracts/core/realm-agent]: Allow the `Net` type parameter to be omitted
By setting a default value. This will allow apps using this class so simply use
`RealmAgent` which is now the same as `RealmAgent<SupportedNetworkName>`. That
way users of this code can work with any supported network in the future, and
not just `'kovan'`.

Also update the example to showcase the intended usage.