build(docker): :building_construction: Implement more deterministic docker image tagging based on git tree hash
Previously, we used to tag the `install` docker image with the SHA1 hash
of the `yarn.lock` file. The idea was that since the `install` image is
used mostly to cache the giant `node_modules` folders, the hash of the
yarn.lock file could be used as a proxy to identify the image contents.

(Alternatively, we could tag the image with the name of the branch (we
did so before that), but that way, since for each new merge request we
have a different branch name, we won't have a corresponding docker
image, even though we could have reused an existing image perfectly
fine, and so the long process of building of the `install` image would
be repeated on every new MR. Even worse, if a commit is added to the
same branch, the old docker image would be reused (since the tag is the
same), which would lead to running other CI jobs in the wrong image.)

However, tagging the image with the SHA1 hash of the `yarn.lock` file
was problematic, since it didn't take into consideration other important
factors that go into determining the image contents. This could lead to
the CI pipeline failing because it spawned the job container into the
wrong `install` image. For example, if in one commit we update the
version of Node.js, but we don't add/remove/update any NPM packages, the
`yarn.lock` file will remain the same and so the jobs in the CI will run
in the old image (with old Node.js).

This commit fixes that by tagging each `install` image with the git tree
hash of a folder containing:

* Select fields of all `package.json` files in the repo (check the
  `JQ_EXPR` defined in `scripts/gather_files_for_yarn_install.bash`)
* The `yarn.lock` file
* The Dockerfile

This way, if say the Dockerfile is modified (e.g. a new package is added to
the base image, or we change the commands run when it's being built), we
would have a different tree hash, which would prevent us from running in
a wrong image. Ditto for the `package.json` files.