config(root): :wrench: Use `direnv` to move all `.env*` files into the root folder
Before, one `.env` / `.env.local` file was needed for each package. This
often led to surprising errors if not all files were updated with the
right values.

With this change, with take advantage of `direnv` support for
automatically sourcing `.env` files (if they exist next to the `.envrc`
file at the repo root folder) and move all settings there.

This commit closes JAR-358.